requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cv==1.0.0
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==19.24.4
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.23.6
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
